Fixed up renaming changes
