+++ package.json
@@ -20,14 +20,15 @@
     "lib:wasm:build:cp": "cpy 'lib/assembly/**/*' dist",
     "lib:js:build": "rollup -c --environment LIB,PROD",
     "lib:js:build:dev": "rollup -c --environment LIB,DEV",
-    "lib:test:build": "asc test/assembly/test.ts -b test/assembly/test.wasm --validate --runtime full --debug",
+    "lib:test:build": "asc lib/assembly/as-bind.ts test/assembly/test.ts -b test/assembly/test.wasm --validate --runtime full --debug",
     "lib:deploy": "np",
     "md:build": "run-s md:wasm:build md:ts:build md:js:build",
-    "md:dev": "run-p lib:watch md:wasm:watch md:js:watch serve",
+    "md:dev": "run-p lib:watch md:wasm:watch md:ts:watch md:js:watch serve",
     "md:wasm:watch": "chokidar --initial \"examples/markdown-parser/assembly/**/*\" -c \"run-s md:wasm:build\"",
+    "md:ts:watch": "chokidar --initial \"examples/markdown-parser/assembly/**/*\" -c \"run-s md:ts:build\"",
     "md:js:watch": "chokidar --initial \"examples/markdown-parser/**/*\" -c \"rollup -c --environment MD,DEV\"",
     "md:wasm:build": "asc lib/assembly/as-bind.ts examples/markdown-parser/assembly/index.ts -b dist/examples/markdown-parser/index.wasm -t dist/examples/markdown-parser/index.wat --sourceMap dist/examples/markdown-parser/index.wasm.map --validate --runtime full",
-    "md:ts:build": "tsc --project examples/markdown-parser/assembly/ --outDir dist/ --module \"es2015\"",
+    "md:ts:build": "tsc --project examples/markdown-parser/assembly/ --outDir dist/ts/ --module \"es2015\"",
     "md:js:build": "rollup -c --environment MD,PROD",
     "md:deploy": "run-s build md:build md:deploy:gh-pages",
     "md:deploy:gh-pages": "gh-pages -d dist/examples/markdown-parser"
